fix: Resolve all golangci-lint issues
- Fix gofmt/goimports formatting issues
- Fix misspell errors across codebase
- Fix unconvert unnecessary type conversions
- Fix bodyclose HTTP response body leaks
- Fix noctx missing context in HTTP requests
- Fix unused code (functions, variables, types)
- Fix unparam unused/constant parameters
- Fix staticcheck issues (deprecated APIs, empty branches, nil contexts)
- Fix govet shadow variable declarations
- Fix gosec security issues (file permissions, command injection, etc.)
- Fix errcheck unchecked error returns

Total: 174 files modified across the codebase.

+++ internal/vpn/conntrack_test.go
@@ -360,7 +360,7 @@ func TestNATTableAllocate(t *testing.T) {
 	require.NotNil(t, entry)
 	assert.Equal(t, src, entry.OriginalSrc)
 	assert.Equal(t, dst, entry.Destination)
-	assert.Equal(t, uint8(ProtocolTCP), entry.Protocol)
+	assert.Equal(t, ProtocolTCP, entry.Protocol)
 	assert.Equal(t, uint16(10000), entry.MappedSrc.Port())
 
 	// Allocate same source/dst should return existing
